+++ Common/Models/Database/Equipment.cs
@@ -4,9 +4,9 @@ namespace FlAdmin.Common.Models.Database;
 
 public class Equipment
 {
-    [BsonElement("equipId")] public long EquipmentId { get; set; }
+    [BsonElement("equipId")] public int EquipmentId { get; set; }
 
     [BsonElement("hardPoint")] public required string HardPoint { get; set; }
 
-    [BsonElement("health")] public long Health { get; set; }
+    [BsonElement("health")] public float Health { get; set; }
 }
